src/requirements.txt
Flask==2.3.3
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
Werkzeug==2.3.7
Jinja2==3.1.2
google-generativeai==0.8.3
python-dotenv==1.0.0